Chapter2-1-1/src/main/java/com/springboot/lvcr/properties/DemoProperties.java
package com.springboot.lvcr.properties;

import org.springframework.beans.factory.annotation.Value;
import org.springframework.stereotype.Component;

@Component
public class DemoProperties {
	@Value("${blog.name}")
	private String name;
	@Value("${blog.title}")
	private String title;
	@Value("${blog.desc}")
	private String desc;
	public String getName() {
		return name;
	}
	public void setName(String name) {
		this.name = name;
	}
	public String getTitle() {
		return title;
	}
	public void setTitle(String title) {
		this.title = title;
	}
	public String getDesc() {
		return desc;
	}
	public void setDesc(String desc) {
		this.desc = desc;
	}
	
	
	
	
	
}
